docker/requirements.txt
Pillow
torch
diffusers
accelerate
transformers